requirements.txt
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.3.0
Pillow==10.3.0
opencv-python==4.9.0
torch==2.4.0.dev20240422
torchvision==0.19.0.dev20240422
matplotlib==3.8.4
seaborn==0.13.2
tqdm==4.65.0
transformers==4.40.2
wandb==0.17.0
